Split WebGl code into multiple modules
