+++ wasm/src/graphics/mod.rs
@@ -1,7 +1,5 @@
-pub mod context;
 mod graphics_api;
-mod render;
+pub mod renderer;
 mod webgl;
-mod webgl_bindings;
 #[doc(inline)]
 pub use graphics_api::GraphicsApi;
