+++ wasm/src/graphics/webgl/shader.rs
@@ -1,0 +1,7 @@
+const VERTEX_SHADER: &str = include_str!("shader/vertex.glsl");
+const FRAGMENT_SHADER: &str = include_str!("shader/fragment.glsl");
+
+enum ShaderType {
+    Vertex,
+    Fragment,
+}
